25   / 1

+++ file28.c
@@ -3,23 +3,23 @@
 int main()
 {
     int luna = 1969;
-    float nascita;
+    int nascita;
     printf("In che anno sei nato: \n");
-    scanf("%f", &nascita);
-    float x = luna - nascita;
-    float y = nascita - luna;
+    scanf("%i", &nascita);
+    int x = luna - nascita;
+    int y = nascita - luna;
     if(x == 0)
     {
         printf("sei nato nell'anno in cui per la prima volta l'uomo è andato sulla Luna\n");
     }
     else if(luna > nascita)
     {
-        printf("Sei nato %f anni prima dell'anno in cui per la prima volta l'uomo è andato sulla Luna\n", x);
+        printf("Sei nato %i anni prima dell'anno in cui per la prima volta l'uomo è andato sulla Luna\n", x);
     }
     else if( luna < nascita)
     {
         
-        printf("Sei nato %f anni dopo l'anno in cui per la prima volta l'uomo è andato sulla Luna\n", y);
+        printf("Sei nato %i anni dopo l'anno in cui per la prima volta l'uomo è andato sulla Luna\n", y);
     }
     else
     {
